Refactorizacion de los mutadores y la politica. Creacion de mutadores para modelos locales. Parametrizacion de la creacion de mutadores.
